Refactor concatTextHtmlAware to StringUtil.
