+++ java/src/jmri/jmrit/symbolicprog/CombinedLocoSelTreePane.java
@@ -28,7 +28,7 @@ import jmri.jmrit.decoderdefn.DecoderIndexFile;
 import jmri.jmrit.progsupport.ProgModeSelector;
 import jmri.jmrit.roster.Roster;
 import jmri.jmrit.roster.RosterEntry;
-import jmri.util.CvUtil;
+import jmri.util.StringUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -317,7 +317,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
             if (verString.equals("")) {
                 return comment;
             } else {
-                return CvUtil.concatTextHtmlAware(comment, " (CV7=" + verString + ")");
+                return StringUtil.concatTextHtmlAware(comment, " (CV7=" + verString + ")");
             }
         }
     }
@@ -343,7 +343,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
     public void resetSelections() {
         Enumeration<TreeNode> e = dRoot.breadthFirstEnumeration();
         while (e.hasMoreElements()) {
-            ((DecoderTreeNode)e.nextElement()).setIdentified(false);
+            ((DecoderTreeNode) e.nextElement()).setIdentified(false);
         }
         setShowMatchedOnly(false);
         selectedPath = new ArrayList<>();
@@ -389,7 +389,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
         // set everybody not identified
         Enumeration<TreeNode> e = dRoot.breadthFirstEnumeration();
         while (e.hasMoreElements()) { // loop over the tree
-            DecoderTreeNode node = ((DecoderTreeNode)e.nextElement());
+            DecoderTreeNode node = ((DecoderTreeNode) e.nextElement());
             node.setIdentified(false);
         }
 
@@ -405,7 +405,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
             String findModel = f.getModel();
 
             while (e.hasMoreElements()) { // loop over the tree & find node
-                DecoderTreeNode node = ((DecoderTreeNode)e.nextElement());
+                DecoderTreeNode node = ((DecoderTreeNode) e.nextElement());
                 // never match show=NO nodes
                 if (node.getShowable() == DecoderFile.Showable.NO) {
                     continue;
@@ -473,7 +473,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
         ArrayList<DecoderTreeNode> selected = new ArrayList<>();
         selectedPath = new ArrayList<>();
         while (e.hasMoreElements()) {
-            DecoderTreeNode node = (DecoderTreeNode)e.nextElement();
+            DecoderTreeNode node = (DecoderTreeNode) e.nextElement();
             if (node.getParent() != null && node.getParent().toString().equals("Root")) {
                 if (node.toString().equals(pMfg)) {
                     TreePath path = new TreePath(node.getPath());
@@ -494,7 +494,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
             Enumeration<TreeNode> es = dRoot.breadthFirstEnumeration();
 
             while (es.hasMoreElements()) {
-                ((DecoderTreeNode)es.nextElement()).setIdentified(true);
+                ((DecoderTreeNode) es.nextElement()).setIdentified(true);
             }
         }
         if (showMatched.isSelected()) {
@@ -544,7 +544,7 @@ public class CombinedLocoSelTreePane extends CombinedLocoSelPane {
         Enumeration<TreeNode> e = dRoot.breadthFirstEnumeration();
 
         while (e.hasMoreElements()) {
-            DecoderTreeNode node = (DecoderTreeNode)e.nextElement();
+            DecoderTreeNode node = (DecoderTreeNode) e.nextElement();
             DecoderTreeNode parentNode = (DecoderTreeNode) node.getParent();
             if (node.toString().equals(modelString)
                     && parentNode.toString().equals(familyString)) {
