+++ java/test/jmri/util/CvUtilTest.java
@@ -247,60 +247,4 @@ public class CvUtilTest {
         Assert.assertEquals("no bits", CvUtil.getMaskDescription("XXXXXXXX"));
     }
 
-    /**
-     * Test of concatTextHtmlAware method, of class CvUtil.
-     */
-    @Test
-    public void testConcatTextHtmlAware() {
-//        System.out.println("concatTextHtmlAware");
-        String baseText = "Some text";
-        String extraText = " and extra stuff";
-
-        String baseTextHtml = "<html>" + baseText + "</html>";
-        String extraTextHtml = "<html>" + extraText + "</html>";
-
-        String baseTextHtmlUpper = "<HTML>" + baseText + "</HTML>";
-        String extraTextHtmlUpper = "<HTML>" + extraText + "</HTML>";
-
-        String expResultNoHtml = baseText + extraText;
-        String expResultHtml = "<html>" + expResultNoHtml + "</html>";
-
-        String result;
-
-        // test null cases
-        Assert.assertEquals(CvUtil.concatTextHtmlAware(null, null), null);
-        Assert.assertEquals(CvUtil.concatTextHtmlAware(baseText, null), baseText);
-        Assert.assertEquals(CvUtil.concatTextHtmlAware(null, extraText), extraText);
-        Assert.assertEquals(CvUtil.concatTextHtmlAware(baseTextHtml, null), baseTextHtml);
-        Assert.assertEquals(CvUtil.concatTextHtmlAware(null, extraTextHtml), extraTextHtml);
-
-        // test with no HTML
-        result = CvUtil.concatTextHtmlAware(baseText, extraText);
-        Assert.assertEquals(expResultNoHtml, result);
-
-        // test with baseText HTML
-        result = CvUtil.concatTextHtmlAware(baseTextHtml, extraText);
-        Assert.assertEquals(expResultHtml, result);
-
-        // test with extraText HTML
-        result = CvUtil.concatTextHtmlAware(baseText, extraTextHtml);
-        Assert.assertEquals(expResultHtml, result);
-
-        // test with both HTML
-        result = CvUtil.concatTextHtmlAware(baseTextHtml, extraTextHtml);
-        Assert.assertEquals(expResultHtml, result);
-
-        // test with baseText uppercase HTML
-        result = CvUtil.concatTextHtmlAware(baseTextHtmlUpper, extraTextHtml);
-        Assert.assertEquals(expResultHtml, result);
-
-        // test with extraText uppercase HTML
-        result = CvUtil.concatTextHtmlAware(baseTextHtml, extraTextHtmlUpper);
-        Assert.assertEquals(expResultHtml, result);
-
-        // test with both uppercase HTML
-        result = CvUtil.concatTextHtmlAware(baseTextHtmlUpper, extraTextHtmlUpper);
-        Assert.assertEquals(expResultHtml, result);
-    }
-
 }
